Neo4jGraphDatabase/llm-vectors-unstructured/CrewAI/requirements.txt
crewai
langchain_google_genai
python_dotenv